Make a mixin class to handle starting and stopping stratisd
This part should be shared.
